IsomorphlyValidationException refactored to standalone JUnite TestCase
